fix: eslint problems
