+++ client/package.json
@@ -9,7 +9,7 @@
     "build-only": "vite build",
     "type-check": "vue-tsc --noEmit",
     "test": "jest",
-    "lint:script": "eslint --fix src/**/*.{js,ts,json,jsx,tsx} --no-error-on-unmatched-pattern",
+    "lint:script": "eslint --fix src/**/*.{js,ts,json,jsx,tsx,vue} --no-error-on-unmatched-pattern",
     "lint:styles": "eslint --fix src/**/*.{css,scss} --no-error-on-unmatched-pattern",
     "format:all": "prettier --write '**/*.{md,ts,html,css,scss,json}'",
     "format": "prettier --write"
