+++ client/src/App.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import Navbar from '@/components/Navbar.vue';
+import Navbar from '@/components/NavbarMenu.vue';
 import { onBeforeMount } from 'vue';
 import { usePermissionsStore } from './store/permissions_store';
 const permissionStore = usePermissionsStore();
