+++ client/src/assets/style.scss
@@ -1,11 +1,11 @@
 @import "./globals.scss";
 
-@import 'vue-select/dist/vue-select.css' ;
+@import 'vue-select/dist/vue-select.css';
 @import 'v-calendar/style.css';
 @import '@vuepic/vue-datepicker/src/VueDatePicker/style/main.scss';
 
 
 body {
-    margin: 0;
-    background-color: #f9f6f6;
+  margin: 0;
+  background-color: #f9f6f6;
 }
