+++ client/src/components/SearchField.vue
@@ -1,27 +1,27 @@
 <template>
-  <input
-    class="search-inp"
-    placeholder="Начните поиск..."
-    v-model="findTeamTxt"
-  />
+    <input
+            class="search-inp"
+            placeholder="Начните поиск..."
+            v-model="findTeamTxt"
+    />
 </template>
 
 <script setup lang="ts">
 import _ from "lodash";
-import { ref, watch } from "vue";
+import {ref, watch} from "vue";
 
 const findTeamTxt = ref("");
 
 const props = defineProps<{
-  handleTimerSearch: Function;
+    handleTimerSearch: (eventTxt: string) => void;
 }>();
 
 const fetchTimer = _.debounce(() => {
-  props.handleTimerSearch(findTeamTxt.value);
+    props.handleTimerSearch(findTeamTxt.value);
 }, 300);
 
 watch(findTeamTxt, () => {
-  fetchTimer();
+    fetchTimer();
 });
 </script>
