+++ client/src/components/SwitchToggle.vue
