+++ client/src/components/TagElem.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script setup lang="ts">
-const props = defineProps<{
+defineProps<{
   text: string;
 }>();
 </script>
