+++ client/src/components/modals/UpdateTeam.ts
@@ -10,6 +10,6 @@ export default class UpdateTeam {
   charterPath = "";
   id_parent = 0;
 
-  fileUstav: any;
-  fileDocument: any;
+  fileUstav: File | undefined;
+  fileDocument: File | undefined;
 }
