+++ client/src/router/index.tsx
@@ -1,5 +1,6 @@
 import { usePermissionsStore } from "@/store/permissions_store";
 import { createRouter, createWebHistory } from "vue-router";
+import type {Permission} from "@/types";
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -8,19 +9,19 @@ const router = createRouter({
     {
       // Авторизация
       path: "/",
-      component: () => import("@/views/events/News.vue"),
+      component: () => import("@/views/events/NewsMain.vue"),
       meta: {},
     },
     {
       // Авторизация
       path: "/login",
-      component: () => import("@/views/Login.vue"),
+      component: () => import("@/views/LoginForm.vue"),
       meta: {},
     },
     {
       // Страница с мероприятиями (кто её news назвал??)
       path: "/news",
-      component: () => import("@/views/events/News.vue"),
+      component: () => import("@/views/events/NewsMain.vue"),
       meta: {},
     },
     {
@@ -41,7 +42,7 @@ const router = createRouter({
       // !Ответственный за направления
       path: "/directions",
 
-      component: () => import("@/views/Directions.vue"),
+      component: () => import("@/views/DirectionsPage.vue"),
       meta: {
         permission: "can view directions",
       },
@@ -50,7 +51,7 @@ const router = createRouter({
       path: "/statistic",
       name: "Statistic",
 
-      component: () => import("@/views/report/Statistic.vue"),
+      component: () => import("@/views/report/StatisticPage.vue"),
       meta: {
         requiresAuth: true,
         permission: "can view directions",
@@ -59,7 +60,7 @@ const router = createRouter({
     {
       path: "/event/:id?",
 
-      component: () => import("@/views/events/Event.vue"),
+      component: () => import("@/views/events/EventMain.vue"),
       meta: {},
     },
     {
@@ -69,18 +70,9 @@ const router = createRouter({
       meta: {},
     },
     {
-      // Страница с личным кабинетом
-      path: "/account",
-
-      component: () => import("@/views/Account.vue"),
-      meta: {
-        // isLoged: true
-      },
-    },
-    {
       path: "/personal/:username",
       name: "Personal",
-      component: () => import("@/views/Personal.vue"),
+      component: () => import("@/views/UserProfile.vue"),
       meta: {
         requiresAuth: true,
       },
@@ -91,7 +83,7 @@ const router = createRouter({
       name: "Requests",
       props: (route) => ({ userId: route.query.user_id }),
 
-      component: () => import("@/views/user/Requests.vue"),
+      component: () => import("@/views/user/EventsRequests.vue"),
       meta: {
         requiresAuth: true,
       },
@@ -103,8 +95,8 @@ const router = createRouter({
 router.beforeEach((to) => {
   const useStore = usePermissionsStore();
 
-  // @ts-ignore
-  if (to.meta.permission && !useStore.can(to.meta.permission)) {
+  const permission = to.meta.permission as Permission
+  if (permission && !useStore.can(permission)) {
     return {
       path: "/login",
       query: { next: to.fullPath },
