+++ client/src/store/enums/enum_event.ts
@@ -1,19 +1,3 @@
-// export enum Direction{
-//   NID = "Научная деятельность",
-//   KTD ="Культурно-массовая деятельность",
-//   UD = "Учебная деятельность",
-//   SD = "Спортивная деятельность",
-//   OD = "Общественная деятельность",
-//   ALL = "Все"
-// }
-
-export enum Level {
-  UNIVERSITY = "Вузовский",
-  CITY = "Городской",
-  REGION = "Региональный",
-  ALL = "Все",
-}
-
 export enum Status {
   CREATED = "Создана",
   ACCEPTED = "Принята",
@@ -27,14 +11,6 @@ export enum Type {
   ALL = "Все",
 }
 
-export enum Direction {
-  NID = "НИД",
-  KTD = "КТД",
-  UD = "УД",
-  SD = "СД",
-  OD = "ОД",
-  ALL = "Все",
-}
 
 export enum DirectionId {
   NID = 6,
