+++ client/src/store/journals_store.ts
@@ -1,21 +1,18 @@
-import { defineStore } from "pinia";
+import {defineStore} from "pinia";
 import axios from "axios";
 
 export const useJournalStore = defineStore("journals", () => {
   // data will be returned as index 0 - is data, index 1 is count
-  async function fetchJournals(teamId: number = -1): Promise<any> {
+  async function fetchJournals(teamId: number = -1) {
     const res = await axios.get("/api/events/journal", {
       params: { team_id: teamId },
     });
-    const data = res.data;
-
-    return data;
+    return res.data;
   }
 
   async function fetchJournalsByUserId(id: number) {
     const res = await axios.get("/api/events/journal/user/" + id);
-    const data = res.data;
-    return data;
+    return res.data;
   }
 
   return {
