--- client/src/views/Account.vue
@@ -1,7 +1,0 @@
-<template>
-  <div class="wrapper-content">Тут личный кабинет</div>
-</template>
-
-<script setup lang="ts"></script>
-
-<style lang="scss" scoped></style>
