+++ client/src/views/report/GraphicsElems.vue
