+++ client/src/views/user/EventsRequests.vue
@@ -20,10 +20,10 @@
 <script setup lang="ts">
 import "@/assets/nav-second.scss"
 import { ref, onBeforeMount } from "vue";
-import { useRoute } from "vue-router";
+// import { useRoute } from "vue-router";
 
-const route = useRoute();
-const idUser = Number(route.query.user_id);
+// const route = useRoute();
+// const idUser = Number(route.query.user_id);
 
 const itemList = [
   { name: "Мероприятия (заявки)" },
